+++ zc-redis-spring-boot-starter/src/main/java/com/zhengcheng/redis/RedisAutoConfiguration.java
@@ -1,6 +1,5 @@
 package com.zhengcheng.redis;
 
-import com.zhengcheng.redis.lock.RedisDistributedLock;
 import com.zhengcheng.redis.properties.CacheManagerProperties;
 import com.zhengcheng.redis.template.RedisBloomFilter;
 import com.zhengcheng.redis.util.RedisObjectSerializer;
@@ -12,7 +11,6 @@ import org.springframework.cache.CacheManager;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.cache.interceptor.KeyGenerator;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Primary;
 import org.springframework.data.redis.cache.RedisCacheConfiguration;
 import org.springframework.data.redis.cache.RedisCacheManager;