+++ zc-redis-spring-boot-starter/src/main/java/com/zhengcheng/redis/annotation/RequestLimit.java
@@ -1,0 +1,29 @@
+package com.zhengcheng.redis.annotation;
+
+import org.springframework.core.Ordered;
+import org.springframework.core.annotation.Order;
+
+import java.lang.annotation.*;
+
+/**
+ * 请求限流
+ *
+ * @author :    zhangquansheng
+ * @date :    2020/5/12 16:11
+ */
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+@Documented
+@Order(Ordered.HIGHEST_PRECEDENCE)
+public @interface RequestLimit {
+
+    /**
+     * 允许访问的最大次数
+     */
+    int count() default Integer.MAX_VALUE;
+
+    /**
+     * 时间段，单位为毫秒，默认值1秒
+     */
+    long time() default 1000;
+}